% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.dataset.json.R
\name{gen.dataset.json}
\alias{gen.dataset.json}
\title{gen.dataset.json}
\usage{
gen.dataset.json(
  dataframe,
  studyOID,
  metaDataVersionOID,
  dataset.name,
  dataset.label,
  pretty.format = FALSE
)
}
\arguments{
\item{dataframe}{Tibble dataframe with variable labels}

\item{studyOID}{Specify Study OID as String}

\item{metaDataVersionOID}{Specify Metadata Version OID as string}

\item{dataset.name}{Specify dataset name as string}

\item{dataset.label}{Specify dataset label as string}

\item{pretty.format}{TRUE for generating a Dataset-Json string with pretty format}
}
\value{
a Dataset-Json string
}
\description{
gen.dataset.json
}
\examples{
\dontrun{
gen.dataset.json(dm, studyOID="cdisc.com/CDISCPILOT01",
                 metaDataVersionOID="MDV.MSGv2.0.SDTMIG.3.3.SDTM.1.7",
                 dataset.name="DM", dataset.label="Demographics")
}
}
