% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.items.R
\name{gen.items}
\alias{gen.items}
\title{gen.items
Make items json that is a part of Dataset-Json}
\usage{
gen.items(dataframe, dataset.name)
}
\arguments{
\item{dataframe}{Tibble dataframe with variable labels}

\item{dataset.name}{Specify dataset name as string}
}
\value{
items data as dataframe
}
\description{
gen.items
Make items json that is a part of Dataset-Json
}
