% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.var.length.R
\name{get.var.length}
\alias{get.var.length}
\title{get.var.length}
\usage{
get.var.length(dataset)
}
\arguments{
\item{dataset}{Tibble dataframe}
}
\value{
integer Variable length
}
\description{
get.var.length
}
