% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.dataset.json.R
\name{read.dataset.json.tibble}
\alias{read.dataset.json.tibble}
\title{read.dataset.json.tibble}
\usage{
read.dataset.json.tibble(dataset_json)
}
\arguments{
\item{dataset_json}{filepath to dataset-json file as string}
}
\value{
tibble
}
\description{
read.dataset.json.tibble
}
\examples{
library(dplyr)
json_file <- RCurl::getURL("https://raw.githubusercontent.com/cdisc-org/DataExchange-DatasetJson/master/examples/sdtm/dm.json")
dm_tibble <- R4DSJSON::read.dataset.json.tibble(json_file)
}
