% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.dataset.json.R
\name{write.dataset.json}
\alias{write.dataset.json}
\title{write.dataset.json
Write Dataset-Json file.}
\usage{
write.dataset.json(
  file,
  dataframe,
  studyOID,
  metaDataVersionOID,
  dataset.name,
  dataset.label,
  pretty.format = FALSE
)
}
\arguments{
\item{file}{file path to dataset-json as string}

\item{dataframe}{Tibble dataframe with variable labels}

\item{studyOID}{Specify Study OID as String}

\item{metaDataVersionOID}{Specify Metadata Version OID as string}

\item{dataset.name}{Specify dataset name as string}

\item{dataset.label}{Specify dataset label as string}

\item{pretty.format}{TRUE for generating a Dataset-Json string with pretty format}
}
\description{
write.dataset.json
Write Dataset-Json file.
}
\examples{
\dontrun{
write.dataset.json(file="dm.json", dataframe=dm, studyOID="cdisc.com/CDISCPILOT01",
                 metaDataVersionOID="MDV.MSGv2.0.SDTMIG.3.3.SDTM.1.7",
                 dataset.name="DM", dataset.label="Demographics")
}
}
