% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dataset.json.R
\name{read.dataset.json}
\alias{read.dataset.json}
\title{read.dataset.json}
\usage{
read.dataset.json(file)
}
\arguments{
\item{file}{file path to dataset-json as string}
}
\value{
tibble dataframe
}
\description{
read.dataset.json
}
\examples{
\dontrun{
dm <- read.dataset.json(file = "dm.json")
}
}
